implemented with unit tests

+++ src/main/kotlin/com/example/demo/common/dto/BetDto.kt
@@ -1,8 +1,10 @@
 package com.example.demo.common.dto
 
+import java.math.BigDecimal
+
 data class BetDto(
-    val stakeAmount: Int,
-    val winAmount: Int,
+    val stakeAmount: BigDecimal,
+    val winAmount: BigDecimal,
     val betValue: Int,
     val actualValue: Int,
 )
