+++ src/main/kotlin/com/example/demo/common/dto/RegistrationResponseDto.kt
@@ -1,5 +1,6 @@
 package com.example.demo.common.dto
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class RegistrationResponseDto(
@@ -7,5 +8,5 @@ data class RegistrationResponseDto(
     val name: String,
     val surname: String,
     val username: String,
-    val walletBalance: Int
+    val walletBalance: BigDecimal
 )
