+++ src/main/kotlin/com/example/demo/common/dto/TransactionDto.kt
@@ -1,8 +1,9 @@
 package com.example.demo.common.dto
 
 import com.example.demo.common.enum.TransactionType
+import java.math.BigDecimal
 
 data class TransactionDto(
-    val amount: Int,
+    val amount: BigDecimal,
     val type: TransactionType,
 )
